html/browsers/browsing-the-web/navigating-across-documents/source/window-open/entry-ok-incumbent-ok-current-navigated.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<title>window.open(): entry ok, incumbent ok, current navigated, and relevant is current</title>
<div id=log></div>
<script>
  var t = async_test(document.title)
  onload = t.step_func(function() {
    var incumbent = self[0],
        current = self[1],
        target = self[2],
        toDo = current.open.bind(current)
    current.frameElement.onload = t.step_func(function() { incumbent.helper(toDo) })
    current.location = "about:blank"
    target.frameElement.onload = t.step_func(() => {
      assert_equals(target.document.referrer, document.URL)
      t.done()
    })
  })
</script>
<iframe src="support/incumbent.html"></iframe>
<iframe src="support/current.html"></iframe>
<iframe name=uniqueish-target></iframe>
